update: Added rss to seperate tool file
